# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GlobeRise Backend API",
    "description": "DeFi and MLM platform backend API with 4-Wallet architecture, investment packages, referral system, and blockchain integration.",
    "version": "1.0.0",
    "contact": {
      "name": "GlobeRise Support",
      "email": "support@globerise.com"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:6969/api",
      "description": "Development server"
    },
    {
      "url": "https://api.globerise.com/api",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/admin/users": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get all users",
        "description": "Get all users with pagination and filters",
        "operationId": "getAllUsers",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": { "type": "integer", "default": 20 }
          },
          {
            "name": "role",
            "in": "query",
            "schema": { "type": "string", "enum": ["USER", "ADMIN", "MODERATOR"] }
          },
          {
            "name": "verified",
            "in": "query",
            "schema": { "type": "boolean" }
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "users": {
                          "type": "array",
                          "items": { "$ref": "#/components/schemas/AdminUser" }
                        },
                        "pagination": { "$ref": "#/components/schemas/Pagination" }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/users/{userId}": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get user by ID",
        "description": "Get user details by ID",
        "operationId": "getUserById",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": { "$ref": "#/components/schemas/AdminUserDetail" }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Admin"],
        "summary": "Delete user",
        "description": "Delete a user",
        "operationId": "deleteUser",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Cannot delete own account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/users/{userId}/role": {
      "put": {
        "tags": ["Admin"],
        "summary": "Assign role",
        "description": "Assign role to user",
        "operationId": "assignRole",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["role"],
                "properties": {
                  "role": {
                    "type": "string",
                    "enum": ["USER", "ADMIN", "MODERATOR"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": { "$ref": "#/components/schemas/User" }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid role or cannot demote self",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/royalty/trigger": {
      "post": {
        "tags": ["Admin"],
        "summary": "Trigger royalty engine",
        "description": "Manually trigger monthly royalty distribution",
        "operationId": "triggerRoyalty",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Royalty engine executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/debug/chain-user/{address}": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get on-chain user info",
        "description": "Debug: Fetch on-chain user info from smart contract",
        "operationId": "getOnChainUser",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "pattern": "^0x[a-fA-F0-9]{40}$" },
            "description": "Wallet address"
          }
        ],
        "responses": {
          "200": {
            "description": "On-chain user info retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string" },
                    "data": { "$ref": "#/components/schemas/OnChainUser" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Blockchain service not connected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string" },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": { "type": "string" },
                "message": { "type": "string" }
              }
            }
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "is_verified": { "type": "boolean" },
          "role": { "$ref": "#/components/schemas/UserRole" },
          "two_factor_enabled": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "Profile": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" },
          "phone": { "type": "string" },
          "avatarUrl": { "type": "string" },
          "address": { "type": "string" },
          "city": { "type": "string" },
          "state": { "type": "string" },
          "zipCode": { "type": "string" },
          "country": { "type": "string" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "totalPages": { "type": "integer" }
        }
      },
      "AdminUser": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string" },
          "role": { "$ref": "#/components/schemas/UserRole" },
          "is_verified": { "type": "boolean" },
          "two_factor_enabled": { "type": "boolean" },
          "rank": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "AdminUserDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/AdminUser" },
          {
            "type": "object",
            "properties": {
              "google_id": { "type": "string" },
              "profile": { "$ref": "#/components/schemas/Profile" },
              "rankHistory": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "rank": { "type": "string" },
                    "achievedAt": { "type": "string", "format": "date-time" },
                    "totalBusiness": { "type": "string" },
                    "strongestLeg": { "type": "string" },
                    "otherLegs": { "type": "string" }
                  }
                }
              }
            }
          }
        ]
      },
...
      "OnChainUser": {
        "type": "object",
        "properties": {
          "address": { "type": "string" },
          "rankName": { "type": "string" },
          "withdrawableBalance": { "type": "string" },
          "isDormant": { "type": "boolean" },
          "investments": { "type": "array", "items": { "type": "object" } },
          "stakes": { "type": "array", "items": { "type": "object" } }
        }
      },
...
      "UserRole": {
        "type": "string",
        "enum": ["USER", "ADMIN", "MODERATOR"]
      },
...
    }
  }
}